Refactor ID types from int to Long pour tout les model

+++ src/main/java/ma/cabinetplus/dao/ConsultationDAO.java
@@ -4,5 +4,5 @@ import ma.cabinetplus.model.Consultation;
 import java.util.List;
 
 public interface ConsultationDAO extends GenericDAO<Consultation, Long> {
-    List<Consultation> trouverParPatient(int patientId);
+    List<Consultation> trouverParPatient(Long patientId);
 }
